+++ tests/io/test_factory.py
@@ -51,7 +51,7 @@ class TestDatasetImporterFactory(unittest.TestCase):
         importer = self.factory.get_importer(source=source)
         self.assertIsInstance(importer, aspecd.io.AdfImporter)
 
-    def test_winsimdat_file_returns_correct_importer(self):
+    def test_niehsdat_file_returns_correct_importer(self):
         source = os.path.join(ROOTPATH, 'testdata', 'Pyrene.dat')
         importer = self.factory.get_importer(source=source)
-        self.assertIsInstance(importer, cwepr.io.WinSimDatImporter)
+        self.assertIsInstance(importer, cwepr.io.NIEHSDatImporter)